+++ CHANGES.rst
@@ -4,7 +4,17 @@ CHANGES
 `Unreleased <https://github.com/fmigneault/aiu/tree/master>`_ (latest)
 ------------------------------------------------------------------------------------
 
-* Nothing yet.
+* Add ``--force-fetch`` option complementary to ``--no-fetch`` to enforce redownload of files if matches are found in
+  the output directory, instead of reusing previously cached results.
+* Add more result file matching combinations to attempt better detection of cached pre-downloaded audio files. Notably,
+  when the ``--prefix-track`` and/or ``--rename-title`` options are omitted, most predefined match patterns where never
+  able to work due to assumed track indices. A combination of ``{artist} - {song}`` patterns allow to match commonly
+  employed file name conventions for song track naming.
+* Fix and improve common mismatches between desired file name patterns and downloaded ones caused by ``youtube_dl``
+  over-sanitizing valid file-system characters that contained unicode or accents. These characters will now be preserved
+  to better retain the original song title, album and artist names using those characters, and as a side effect, more
+  efficiently match the expected downlaod file name against target ID3 metadata.
+* Add DevOps ``Makefile`` along with multiple inspection, linting and valiation utilities.
 
 `1.9.1 <https://github.com/fmigneault/aiu/tree/1.9.1>`_ (2022-10-30)
 ------------------------------------------------------------------------------------